Display allocation information on combined case check and submit page
