+++ packages/webapp/src/routes/combined-case/check-and-submit.js
@@ -2,8 +2,10 @@ import constants from '../../utils/constants.js'
 import { getTaskList } from '../../journey-validation/task-list-generator.js'
 import application from '../../utils/combined-case-application.js'
 import getRegistrationDetails from '../../utils/get-land-check-and-submit-details.js'
+import getDeveloperDetails from '../../utils/get-developer-check-and-submit-details.js'
 import combinedCaseApplicationValidation from '../../utils/combined-case-application-validation.js'
 import { postJson } from '../../utils/http.js'
+import getOrganisationDetails from '../../utils/get-organisation-details.js'
 
 const handlers = {
   get: (request, h) => {
@@ -21,19 +23,33 @@ const handlers = {
 
     const applicationDetails = application(request.yar, request.auth.credentials.account).combinedCase
     console.log(JSON.stringify(applicationDetails, null, 2))
-
-    return h.view(constants.views.COMBINED_CASE_CHECK_AND_SUBMIT, getRegistrationDetails(request, applicationDetails))
+    const claims = request.auth.credentials.account.idTokenClaims
+    const { currentOrganisation } = getOrganisationDetails(claims)
+    return h.view(
+      constants.views.COMBINED_CASE_CHECK_AND_SUBMIT,
+      {
+        ...getRegistrationDetails(request, applicationDetails),
+        ...getDeveloperDetails(request, request.yar, currentOrganisation)
+      }
+    )
   },
   post: async (request, h) => {
     const combinedCaseApplication = application(request.yar, request.auth.credentials.account)
     const applicationDetails = combinedCaseApplication.combinedCase
 
     if (request.payload.termsAndConditionsConfirmed !== 'Yes') {
+      const claims = request.auth.credentials.account.idTokenClaims
+      const { currentOrganisation } = getOrganisationDetails(claims)
       const err = [{
         text: 'You must confirm you have read the terms and conditions',
         href: '#termsAndConditionsConfirmed'
       }]
-      return h.view(constants.views.COMBINED_CASE_CHECK_AND_SUBMIT, { ...getRegistrationDetails(request, applicationDetails), err })
+      return h.view(constants.views.COMBINED_CASE_CHECK_AND_SUBMIT,
+        {
+          ...getRegistrationDetails(request, applicationDetails),
+          ...getDeveloperDetails(request, request.yar, currentOrganisation),
+          err
+        })
     }
 
     const { value, error } = combinedCaseApplicationValidation.validate(combinedCaseApplication)
